test: cover calcularMensual
