+++ static/css/styles.css
@@ -63,7 +63,7 @@ h3 {
   border-radius: 0 !important;
 }
 .accordion-button {
-  background-color: #1A1A1A !important; /* Rojo Bootstrap danger */
+  background-color: #1A1A1A !important; /* Negro institucional */
   color: white;
   font-weight: bold;
 }
