+++ static/index.html
@@ -27,6 +27,9 @@
     <!-- Bootstrap JS BUNDLE (incluye Popper) -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.7/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
 
+    <!-- Utilidades -->
+    <script src="/static/js/calcularMensual.js"></script>
+
     <!--JS principal (después del bundle) -->
     <script src="/static/js/main.js"></script>
   </body>
